adding cache redis and pait currency

+++ src/favorites/favorites.controller.ts
@@ -1,34 +1,33 @@
-import { Controller, Post, Get, Delete, Body, UseGuards, Request } from '@nestjs/common';
+import { Controller, Post, Get, Delete, Body, UseGuards, Request, Req, UnauthorizedException } from '@nestjs/common';
 import { FavoritesService } from './favorites.service';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
+import { AuthGuard } from '@nestjs/passport';
+import { Request as ExpressRequest } from 'express';
 
 @Controller('favorites')
-@UseGuards(JwtAuthGuard)
 export class FavoritesController {
   constructor(private readonly favoritesService: FavoritesService) {}
 
+  @UseGuards(AuthGuard('jwt'))
   @Post()
-  async addFavoritePair(
-    @Body() body: { fromCurrency: string; toCurrency: string },
-    @Request() req,
-  ) {
-    return this.favoritesService.addFavoritePair(
-      req.user.id,
-      body.fromCurrency,
-      body.toCurrency,
-    );
+  async addFavorite(@Body() body: { fromCurrency: string; toCurrency: string }, @Req() req: ExpressRequest) {
+    if (!req.user) throw new UnauthorizedException();
+    return this.favoritesService.addFavorite(req.user.id, body.fromCurrency, body.toCurrency);
   }
 
+  @UseGuards(AuthGuard('jwt'))
   @Get()
-  async getFavoritePairs(@Request() req) {
-    return this.favoritesService.getFavoritePairs(req.user.id);
+  async getFavorites(@Req() req: ExpressRequest) {
+    if (!req.user) throw new UnauthorizedException();
+    return this.favoritesService.getFavorites(req.user.id);
   }
 
   @Delete()
   async removeFavoritePair(
     @Body() body: { fromCurrency: string; toCurrency: string },
-    @Request() req,
+    @Request() req: ExpressRequest,
   ) {
+    if (!req.user) throw new UnauthorizedException();
     return this.favoritesService.removeFavoritePair(
       req.user.id,
       body.fromCurrency,
